ran './dev-tools/scripts/branchSplitter.py 56842452026' script

+++ solr/core/src/java/org/apache/solr/handler/component/RangeFacetRequest.java
@@ -700,7 +700,7 @@ public class RangeFacetRequest extends FacetComponent.FacetBase {
 
     @Override
     public Float parseAndAddGap(Float value, String gap) {
-      return value + Float.parseFloat(gap);
+      return value.floatValue() + Float.parseFloat(gap);
     }
   }
 
@@ -717,7 +717,7 @@ public class RangeFacetRequest extends FacetComponent.FacetBase {
 
     @Override
     public Double parseAndAddGap(Double value, String gap) {
-      return value + Double.parseDouble(gap);
+      return value.doubleValue() + Double.parseDouble(gap);
     }
   }
 
@@ -734,7 +734,7 @@ public class RangeFacetRequest extends FacetComponent.FacetBase {
 
     @Override
     public Integer parseAndAddGap(Integer value, String gap) {
-      return value + Integer.parseInt(gap);
+      return value.intValue() + Integer.parseInt(gap);
     }
   }
 
@@ -751,7 +751,7 @@ public class RangeFacetRequest extends FacetComponent.FacetBase {
 
     @Override
     public Long parseAndAddGap(Long value, String gap) {
-      return value + Long.parseLong(gap);
+      return value.longValue() + Long.parseLong(gap);
     }
   }
