+++ solr/core/src/java/org/apache/solr/schema/RandomSortField.java
@@ -129,7 +129,7 @@ public class RandomSortField extends FieldType {
 
             @Override
             public void setTopValue(Integer value) {
-              topVal = value;
+              topVal = value.intValue();
             }
 
             @Override
