+++ solr/core/src/java/org/apache/solr/uninverting/FieldCacheImpl.java
@@ -957,7 +957,7 @@ public class FieldCacheImpl implements FieldCache {
 
       Terms terms = reader.terms(key.field);
 
-      final float acceptableOverheadRatio = (Float) key.custom;
+      final float acceptableOverheadRatio = ((Float) key.custom).floatValue();
 
       final PagedBytes bytes = new PagedBytes(15);
 
@@ -1171,7 +1171,7 @@ public class FieldCacheImpl implements FieldCache {
       final int maxDoc = reader.maxDoc();
       Terms terms = reader.terms(key.field);
 
-      final float acceptableOverheadRatio = (Float) key.custom;
+      final float acceptableOverheadRatio = ((Float) key.custom).floatValue();
 
       final int termCountHardLimit = maxDoc;
