+++ solr/core/src/test/org/apache/solr/search/AnalyticsTestQParserPlugin.java
@@ -124,7 +124,7 @@ public class AnalyticsTestQParserPlugin extends QParserPlugin {
         @SuppressWarnings({"rawtypes"})
         NamedList analytics = (NamedList) response.get("analytics");
         Integer c = (Integer) analytics.get("mycount");
-        count += c;
+        count += c.intValue();
       }
 
       merged.add("mycount", count);
@@ -143,7 +143,7 @@ public class AnalyticsTestQParserPlugin extends QParserPlugin {
         @SuppressWarnings({"rawtypes"})
         NamedList analytics = (NamedList) response.get("analytics");
         Integer c = (Integer) analytics.get("mycount");
-        count += c;
+        count += c.intValue();
       }
 
       ModifiableSolrParams params = new ModifiableSolrParams();
@@ -165,7 +165,7 @@ public class AnalyticsTestQParserPlugin extends QParserPlugin {
         @SuppressWarnings({"rawtypes"})
         NamedList analytics = (NamedList) response.getResponse().get("analytics");
         Integer c = (Integer) analytics.get("mycount");
-        nextCount += c;
+        nextCount += c.intValue();
       }
 
       @SuppressWarnings({"rawtypes"})
