+++ solr/core/src/test/org/apache/solr/search/SolrIndexSearcherTest.java
@@ -273,7 +273,7 @@ public class SolrIndexSearcherTest extends SolrTestCaseJ4 {
               if (fixedScore == null) {
                 expectedScore = 1f;
               } else {
-                expectedScore = fixedScore;
+                expectedScore = fixedScore.floatValue();
                 cmd.setQuery(new FixedScoreReRankQuery(cmd.getQuery(), expectedScore));
               }
