+++ solr/solrj/src/java/org/apache/solr/common/util/FastJavaBinDecoder.java
@@ -312,7 +312,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return (byte) entry.numericVal;
+        return Byte.valueOf((byte) entry.numericVal);
       }
     },
     _SHORT(SHORT, LOWER_5_BITS, DataEntry.Type.INT) {
@@ -324,7 +324,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return (short) entry.numericVal;
+        return Short.valueOf((short) entry.numericVal);
       }
     },
     _DOUBLE(DOUBLE, LOWER_5_BITS, DataEntry.Type.DOUBLE) {
@@ -336,7 +336,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return entry.doubleVal;
+        return Double.valueOf(entry.doubleVal);
       }
     },
     _INT(INT, LOWER_5_BITS, DataEntry.Type.INT) {
@@ -347,7 +347,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return (int) entry.numericVal;
+        return Integer.valueOf((int) entry.numericVal);
       }
     }, // signed integer
     _LONG(LONG, LOWER_5_BITS, DataEntry.Type.LONG) {
@@ -358,7 +358,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return entry.numericVal;
+        return Long.valueOf(entry.numericVal);
       }
     },
     _FLOAT(FLOAT, LOWER_5_BITS, DataEntry.Type.FLOAT) {
@@ -369,7 +369,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return (float) entry.doubleVal;
+        return Float.valueOf((float) entry.doubleVal);
       }
     },
     _DATE(DATE, LOWER_5_BITS, DataEntry.Type.DATE) {
@@ -583,7 +583,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return (int) entry.numericVal;
+        return Integer.valueOf((int) entry.numericVal);
       }
     },
     _SLONG(SLONG, UPPER_3_BITS, DataEntry.Type.LONG) {
@@ -594,7 +594,7 @@ public class FastJavaBinDecoder implements DataEntry.FastDecoder {
 
       @Override
       public Object readObject(StreamCodec codec, EntryImpl entry) {
-        return (long) (int) entry.numericVal;
+        return Long.valueOf((int) entry.numericVal);
       }
     },
     _ARR(ARR, UPPER_3_BITS, DataEntry.Type.ENTRY_ITER) {
